requirements.txt
PyQt6
websocket-client
requests
